// resume-api/src/user-management/models/add-user-contact.view-model.ts
/**
 * Represents an add user contact view model
 */
export interface AddUserContactViewModel {
    
    /**
     * Gets or sets the user identifier
     */
    userId: string;

    /**
     * Gets or sets the contact type
     */
    contactType: string;

    /**
     * Gets or sets the contact value
     */
    contactValue: string;
}