// resume-api/src/user-management/models/add-user-langauge.view-model.ts
/**
 * Represents the add user language view model
 */
export interface AddUserLanguageViewModel {
    /**
     * Gets or sets the user identifier
     */
    userId: string;
    
    /**
     * Gets or sets the language
     */
    language: string;

    /**
     * Gets or sets the leval;
     */
    level: number;
}